+++ flash-api/src/main/java/cn/enilu/flash/api/controller/system/CfgController.java
@@ -3,7 +3,6 @@ package cn.enilu.flash.api.controller.system;
 import cn.enilu.flash.api.controller.BaseController;
 import cn.enilu.flash.bean.constant.factory.PageFactory;
 import cn.enilu.flash.bean.core.BussinessLog;
-import cn.enilu.flash.bean.dictmap.CfgDict;
 import cn.enilu.flash.bean.entity.system.Cfg;
 import cn.enilu.flash.bean.entity.system.FileInfo;
 import cn.enilu.flash.bean.enumeration.BizExceptionEnum;
@@ -43,14 +42,14 @@ public class CfgController extends BaseController {
     /**
      * 查询参数列表
      */
-    @RequestMapping(value = "/list",method = RequestMethod.GET)
-    @RequiresPermissions(value = {Permission.CFG})
+    @GetMapping(value = "/list")
+    @RequiresPermissions(value = {"/cfg"})
     public Object list(@RequestParam(required = false) String cfgName, @RequestParam(required = false) String cfgValue) {
         Page<Cfg> page = new PageFactory<Cfg>().defaultPage();
-        if(StringUtil.isNotEmpty(cfgName)){
+        if (StringUtil.isNotEmpty(cfgName)) {
             page.addFilter(SearchFilter.build("cfgName", SearchFilter.Operator.LIKE, cfgName));
         }
-        if(StringUtil.isNotEmpty(cfgValue)){
+        if (StringUtil.isNotEmpty(cfgValue)) {
             page.addFilter(SearchFilter.build("cfgValue", SearchFilter.Operator.LIKE, cfgValue));
         }
         page = cfgService.queryPage(page);
@@ -59,47 +58,57 @@ public class CfgController extends BaseController {
 
     /**
      * 导出参数列表
+     *
      * @param cfgName
      * @param cfgValue
      * @return
      */
-    @RequestMapping(value = "/export",method = RequestMethod.GET)
+    @GetMapping(value = "/export")
     @RequiresPermissions(value = {Permission.CFG})
     public Object export(@RequestParam(required = false) String cfgName, @RequestParam(required = false) String cfgValue) {
         Page<Cfg> page = new PageFactory<Cfg>().defaultPage();
-        if(StringUtil.isNotEmpty(cfgName)){
+        if (StringUtil.isNotEmpty(cfgName)) {
             page.addFilter(SearchFilter.build("cfgName", SearchFilter.Operator.LIKE, cfgName));
         }
-        if(StringUtil.isNotEmpty(cfgValue)){
+        if (StringUtil.isNotEmpty(cfgValue)) {
             page.addFilter(SearchFilter.build("cfgValue", SearchFilter.Operator.LIKE, cfgValue));
         }
         page = cfgService.queryPage(page);
-        FileInfo fileInfo = fileService.createExcel("templates/config.xlsx","系统参数.xlsx",Maps.newHashMap("list",page.getRecords()));
+        FileInfo fileInfo = fileService.createExcel("templates/config.xlsx", "系统参数.xlsx", Maps.newHashMap("list", page.getRecords()));
         return Rets.success(fileInfo);
     }
-    @RequestMapping(method = RequestMethod.POST)
-    @BussinessLog(value = "编辑参数", key = "cfgName",dict= CfgDict.class)
-    @RequiresPermissions(value = {Permission.CFG_EDIT})
-    public Object save(@ModelAttribute @Valid Cfg cfg){
-        if(cfg.getId()!=null){
-            Cfg old = cfgService.get(cfg.getId());
-            LogObjectHolder.me().set(old);
-            old.setCfgName(cfg.getCfgName());
-            old.setCfgValue(cfg.getCfgValue());
-            old.setCfgDesc(cfg.getCfgDesc());
-            cfgService.saveOrUpdate(old);
-        }else {
-            cfgService.saveOrUpdate(cfg);
-        }
+
+    @PostMapping
+    @BussinessLog(value = "新增参数", key = "cfgName")
+    @RequiresPermissions(value = {"/cfg/add"})
+    public Object add(@RequestBody @Valid Cfg cfg) {
+        cfgService.saveOrUpdate(cfg);
         return Rets.success();
     }
-    @RequestMapping(method = RequestMethod.DELETE)
-    @BussinessLog(value = "删除参数", key = "id",     dict= CfgDict.class)
-    @RequiresPermissions(value = {Permission.CFG_DEL})
-    public Object remove(@RequestParam Long id){
-        logger.info("id:{}",id);
+
+    @PutMapping
+    @BussinessLog(value = "编辑参数", key = "cfgName")
+    @RequiresPermissions(value = {"/cfg/update"})
+    public Object update(@RequestBody @Valid Cfg cfg) {
+        Cfg old = cfgService.get(cfg.getId());
+        LogObjectHolder.me().set(old);
+        old.setCfgName(cfg.getCfgName());
+        old.setCfgValue(cfg.getCfgValue());
+        old.setCfgDesc(cfg.getCfgDesc());
+        cfgService.saveOrUpdate(old);
+        return Rets.success();
+    }
+
+    @DeleteMapping
+    @BussinessLog(value = "删除参数", key = "id")
+    @RequiresPermissions(value = {"/cfg/delete"})
+    public Object remove(@RequestParam Long id) {
+        logger.info("id:{}", id);
         if (id == null) {
             throw new ApplicationException(BizExceptionEnum.REQUEST_NULL);
+        }
+        if(id<9){
+            return Rets.failure("禁止删除初始化参数");
         }
         cfgService.delete(id);
         return Rets.success();
