+++ flash-api/src/main/java/cn/enilu/flash/api/controller/system/DeptContoller.java
@@ -2,7 +2,6 @@ package cn.enilu.flash.api.controller.system;
 
 import cn.enilu.flash.api.controller.BaseController;
 import cn.enilu.flash.bean.core.BussinessLog;
-import cn.enilu.flash.bean.dictmap.DeptDict;
 import cn.enilu.flash.bean.entity.system.Dept;
 import cn.enilu.flash.bean.enumeration.BizExceptionEnum;
 import cn.enilu.flash.bean.enumeration.Permission;
@@ -34,42 +33,46 @@ public class DeptContoller extends BaseController {
 
     @Autowired
     private DeptService deptService;
-    @RequestMapping(value = "/list",method = RequestMethod.GET)
+
+    @GetMapping(value = "/list")
     @RequiresPermissions(value = {Permission.DEPT})
-    public Object list(){
+    public Object list() {
         List<DeptNode> list = deptService.queryAllNode();
         return Rets.success(list);
     }
-    @RequestMapping(method = RequestMethod.POST)
-    @BussinessLog(value = "编辑部门", key = "simplename", dict = DeptDict.class)
+
+    @PostMapping
+    @BussinessLog(value = "编辑部门", key = "simplename")
     @RequiresPermissions(value = {Permission.DEPT_EDIT})
-    public Object save(@ModelAttribute @Valid Dept dept){
+    public Object save(@RequestBody @Valid Dept dept) {
         if (BeanUtil.isOneEmpty(dept, dept.getSimplename())) {
             throw new ApplicationException(BizExceptionEnum.REQUEST_NULL);
         }
-        if(dept.getId()!=null){
+        if (dept.getId() != null) {
             Dept old = deptService.get(dept.getId());
             LogObjectHolder.me().set(old);
             old.setPid(dept.getPid());
             old.setSimplename(dept.getSimplename());
             old.setFullname(dept.getFullname());
             old.setNum(dept.getNum());
-            old.setTips(dept.getTips());
             deptService.deptSetPids(old);
             deptService.update(old);
-        }else {
+        } else {
             deptService.deptSetPids(dept);
             deptService.insert(dept);
         }
         return Rets.success();
     }
-    @RequestMapping(method = RequestMethod.DELETE)
-    @BussinessLog(value = "删除部门", key = "id", dict = DeptDict.class)
+
+    @DeleteMapping
+    @BussinessLog(value = "删除部门", key = "id")
     @RequiresPermissions(value = {Permission.DEPT_DEL})
-    public Object remove(@RequestParam  Long id){
-        logger.info("id:{}",id);
+    public Object remove(@RequestParam Long id) {
         if (id == null) {
             throw new ApplicationException(BizExceptionEnum.REQUEST_NULL);
+        }
+        if(id<5){
+            return Rets.failure("禁止删除初始部门");
         }
         deptService.deleteDept(id);
         return Rets.success();
